widescreen

+++ src/game/print.h
@@ -25,9 +25,9 @@
 #define GLYPH_DOUBLE_QUOTE    57
 #define GLYPH_UMLAUT          58
 
-void print_text_fmt_int(s32 x, s32 y, const char *str, s32 n);
-void print_text(s32 x, s32 y, const char *str);
-void print_text_centered(s32 x, s32 y, const char *str);
+void print_text_fmt_int(f32 x, f32 y, const char *str, s32 n);
+void print_text(f32 x, f32 y, const char *str);
+void print_text_centered(f32 x, f32 y, const char *str);
 void render_text_labels(void);
 
 #endif // PRINT_H
